requirements.txt
arxiv==2.1.0
PyYAML==6.0.1